backend/requirements.txt
asgiref==3.8.1
Django==5.1.4
sqlparse==0.5.3
opal-client==0.1.6
djangorestframework==3.15.2
wdb==3.3.0
django-cors-headers==4.6.0
psycopg2==2.9.10
djangorestframework-simplejwt==5.3.1